# PERF: extract_array