--- a/pandas/core/base.py
+++ b/pandas/core/base.py
@@ -907,17 +907,12 @@ class IndexOpsMixin(OpsMixin):
             If the function returns a tuple with more than one element
             a MultiIndex will be returned.
         """
-        arr = extract_array(self, extract_numpy=True, extract_range=True)
+        arr = self._values
 
         if isinstance(arr, ExtensionArray):
             return arr.map(mapper, na_action=na_action)
 
-        # Argument 1 to "map_array" has incompatible type
-        # "Union[IndexOpsMixin, ndarray[Any, Any]]";
-        # expected "Union[ExtensionArray, ndarray[Any, Any]]
-        return algorithms.map_array(
-            arr, mapper, na_action=na_action, convert=convert  # type: ignore[arg-type]
-        )
+        return algorithms.map_array(arr, mapper, na_action=na_action, convert=convert)
 
     @final
     def value_counts(