```
import numpy as np
import pandas as pd

arr = np.arange(5)
ser = pd.Series(arr)

%timeit pd.core.construction.extract_array(arr, extract_numpy=True)
924 ns ± 23.5 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # <- main
205 ns ± 15.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # <- PR

%timeit pd.core.construction.extract_array(ser, extract_numpy=True)
1.34 µs ± 20.8 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # <- main
471 ns ± 9.53 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # <- PR
```